Adds HotSigner module
